backend/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.2.6
django-rest-framework==0.1.0
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.4
numpy==1.21.4
packaging==23.2
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.6
